Generics improvements:
automatically sub-class Structured generics in cases where it won't
interfere with how `typing.Generic` normally behaves. So this is
limited to when used as a type-hint within a non-generic `Structured`
class.  This is nested to arbitrary levels, so only the outer-most
`Structured` class needs to be non-generic.

Otherwise, manual sub-classing will still work to get the concrete
implementation of the class.